# docs/openapi.yaml
openapi: 3.1.0
info:
  title: StayHub API
  version: 0.1.0
  description: APISIX-facing API contract for the StayHub hotel reservation platform.
servers:
  - url: http://localhost:9080
    description: Local APISIX gateway
security:
  - bearerAuth: []
...
paths:
...
  /api/inventory/reservations:
    post:
      tags: [Inventory]
      summary: Reserve room inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryReservationInput'
      responses:
        '200':
          description: Inventory reserved
        '409':
          $ref: '#/components/responses/Conflict'
  /api/inventory/releases:
    post:
      tags: [Inventory]
      summary: Release room inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryReservationInput'
      responses:
        '200':
          description: Inventory released
  /api/bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      description: Customers see only their own bookings. Managers and admins can list all bookings.
      parameters:
        - name: hotel_id
          in: query
          schema:
            type: integer
        - name: room_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated bookings
    post:
      tags: [Bookings]
      summary: Create booking
      description: Derives the user from the authenticated gateway identity, reserves inventory, calculates the total from inventory pricing, creates a pending payment, and creates a notification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingInput'
      responses:
        '201':
          description: Booking created
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/bookings/{booking}:
    parameters:
      - $ref: '#/components/parameters/BookingId'
    get:
      tags: [Bookings]
      summary: Show booking
      responses:
        '200':
          description: Booking
...
  /api/payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - name: booking_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated payments
    post:
      tags: [Payments]
      summary: Create mock payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          description: Payment created
...
  /api/payments/{payment}/succeed:
    post:
      tags: [Payments]
      summary: Mark pending payment succeeded
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment succeeded
  /api/payments/{payment}/fail:
    post:
      tags: [Payments]
      summary: Mark pending payment failed
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                failure_reason:
                  type: string
      responses:
        '200':
          description: Payment failed
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters:
        - name: recipient_user_id
          in: query
          schema:
            type: integer
        - name: channel
          in: query
          schema:
            type: string
            enum: [email, sms, in_app]
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: unread
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated notifications
    post:
      tags: [Notifications]
      summary: Create notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationInput'
      responses:
        '201':
          description: Notification created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token from the `stayhub` realm.
  parameters:
...
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
...
    BookingId:
      name: booking
      in: path
      required: true
      schema:
        type: integer
    PaymentId:
      name: payment
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, cancelled, payment_failed]
    PaymentStatus:
      type: string
      enum: [pending, succeeded, failed, refunded]
    NotificationStatus:
      type: string
      enum: [queued, sent, failed]
...
    InventoryReservationInput:
      type: object
      required: [room_id, check_in, check_out]
      properties:
        room_id:
          type: integer
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        quantity:
          type: integer
          default: 1
    BookingInput:
      type: object
      required: [hotel_id, room_id, check_in, check_out]
      properties:
        hotel_id:
          type: integer
        room_id:
          type: integer
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        quantity:
          type: integer
          default: 1
    PaymentInput:
      type: object
      required: [booking_id, user_id, amount, currency]
      properties:
        booking_id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          example: USD
        provider:
          type: string
          example: mock
...
    NotificationInput:
      type: object
      required: [recipient_user_id, channel, type, subject, body]
      properties:
        recipient_user_id:
          type: integer
        channel:
          type: string
          enum: [email, sms, in_app]
        type:
          type: string
        subject:
          type: string
        body:
          type: string
        payload:
          type: object
...
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object